+++ app/src/androidTest/java/woowacourse/movie/MovieReservationActivityTest.kt
@@ -16,10 +16,10 @@ import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
 import woowacourse.movie.constants.MovieContentKey
-import woowacourse.movie.data.MovieContentsImpl
 import woowacourse.movie.model.Date
 import woowacourse.movie.model.MovieContent
-import woowacourse.movie.view.MovieReservationActivity
+import woowacourse.movie.model.data.MovieContentsImpl
+import woowacourse.movie.view.activity.MovieReservationActivity
 
 @RunWith(AndroidJUnit4::class)
 class MovieReservationActivityTest {
