+++ app/src/androidTest/java/woowacourse/movie/MovieReservationCompleteActivityTest.kt
@@ -15,10 +15,10 @@ import org.junit.Test
 import org.junit.runner.RunWith
 import woowacourse.movie.constants.MovieContentKey
 import woowacourse.movie.constants.MovieReservationKey
-import woowacourse.movie.data.MovieContentsImpl
 import woowacourse.movie.model.Date
 import woowacourse.movie.model.MovieContent
-import woowacourse.movie.view.MovieReservationCompleteActivity
+import woowacourse.movie.model.data.MovieContentsImpl
+import woowacourse.movie.view.activity.MovieReservationCompleteActivity
 
 @RunWith(AndroidJUnit4::class)
 class MovieReservationCompleteActivityTest {
