+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.Movie"
         tools:targetApi="31">
         <activity
-            android:name=".view.MainActivity"
+            android:name=".view.activity.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -22,10 +22,10 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".view.MovieReservationActivity"
+            android:name=".view.activity.MovieReservationActivity"
             android:exported="false" />
         <activity
-            android:name=".view.MovieReservationCompleteActivity"
+            android:name=".view.activity.MovieReservationCompleteActivity"
             android:exported="false" />
     </application>
