+++ app/src/main/java/woowacourse/movie/model/data/MovieContents.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.data
+package woowacourse.movie.model.data
 
 import woowacourse.movie.model.MovieContent
